Split endpoints into different files
