+++ tests/conftest.py
@@ -4,7 +4,7 @@ import pytest
 import werkzeug
 from fastapi.testclient import TestClient
 from src import database
-from src.api.nextcloud_news.endpoint import app
+from src.api.nextcloud_news.app import app
 
 
 @pytest.fixture
